Move pipeline assertions to support folder
To stop it being picked up as a spec file
